fix: normalize buttons
