+++ src/plugins/queryhistory/components/CloseButton/CloseButton.js
@@ -1,0 +1,11 @@
+import { StyledCloseButton } from "../../styled";
+import CloseIcon from "@mui/icons-material/Close";
+
+
+export default function CloseButton({ onClose }) {
+    return (
+        <StyledCloseButton onClick={onClose}>
+            <CloseIcon />
+        </StyledCloseButton>
+    );
+}
