+++ src/plugins/queryhistory/components/EmptyHistoryDisplay/EmptyHistoryDisplay.js
@@ -1,0 +1,5 @@
+import { EmptyHistory } from "../../styled";
+
+export default function EmptyHistoryDisplay({ message }) {
+    return <EmptyHistory>{message}</EmptyHistory>;
+}
