+++ src/plugins/queryhistory/index.js
@@ -1,5 +1,4 @@
-import styled from "@emotion/styled";
-import { createTheme, ThemeProvider, Tooltip } from "@mui/material";
+import { ThemeProvider, Tooltip } from "@mui/material";
 import localService from "../../services/localService";
 import Drawer from "@mui/material/Drawer";
 import { useState, useEffect } from "react";
@@ -13,197 +12,45 @@ import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
 import StarIcon from "@mui/icons-material/Star";
 import StarBorderIcon from "@mui/icons-material/StarBorder";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
-import DisplaySettingsIcon from "@mui/icons-material/DisplaySettings";
-import HistoryIcon from "@mui/icons-material/History";
-import CloseIcon from "@mui/icons-material/Close";
-import SearchIcon from "@mui/icons-material/Search";
+
 import LinkIcon from "@mui/icons-material/Link";
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
 import TabsUnstyled from "@mui/base/TabsUnstyled";
-import TabsListUnstyled from "@mui/base/TabsListUnstyled";
-import TabPanelUnstyled from "@mui/base/TabPanelUnstyled";
-import { buttonUnstyledClasses } from "@mui/base/ButtonUnstyled";
-import TabUnstyled, { tabUnstyledClasses } from "@mui/base/TabUnstyled";
+
 import LaunchIcon from "@mui/icons-material/Launch";
 import KeyboardArrowRightIcon from "@mui/icons-material/KeyboardArrowRight";
 import ZoomOutMapIcon from "@mui/icons-material/ZoomOutMap";
 import ZoomInMapIcon from "@mui/icons-material/ZoomInMap";
-// Dialog
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogContentText from "@mui/material/DialogContentText";
-import DialogTitle from "@mui/material/DialogTitle";
+
 import { notificationTypes } from "../notifications/consts";
 import localUrl from "../../services/localUrl";
 
 import setLinksHistory from "../../actions/setLinksHistory";
-//import { DateRangePickerMain } from "../../components/StatusBar/components/daterangepicker/index";
-// Alert Dialog for Clearing History
-function AlertDialog({ clearHistory, dialogType }) {
-    const [open, setOpen] = useState(false);
+import {
+    TabsList,
+    Tab,
+    TabHistoryIcon,
+    TabHistoryLinkIcon,
+    TabHistoryStarIcon,
+    TabHistorySettingIcon,
+    TabPanel,
+    LinkParams,
+    RowData,
+    HistoryButton,
+    SubmitButton,
+    QueryHistoryContainer,
+    TabHeaderContainer,
+    TabHistorySearchIcon,
+    FilterInput,
+    SettingItemContainer,
+    TimeSpan,
+    HistoryRow,
+    theme,
+} from "./styled";
 
-    const handleClickOpen = () => {
-        setOpen(true);
-    };
-
-    const handleClose = () => {
-        setOpen(false);
-    };
-    function handleClearHistory() {
-        clearHistory();
-        setOpen(false);
-    }
-    return (
-        <div>
-            <Tooltip title={"Clear Query History"}>
-                <ClearHistoryButton onClick={handleClickOpen}>
-                    {"Clear History"}
-                </ClearHistoryButton>
-            </Tooltip>
-            <Dialog
-                open={open}
-                onClose={handleClose}
-                aria-labelledby="alert-dialog-title"
-                aria-describedby="alert-dialog-description"
-            >
-                <DialogTitle id="alert-dialog-title">
-                    Are you sure you want to clear the {dialogType} History?
-                </DialogTitle>
-                <DialogContent>
-                    <DialogContentText id="alert-dialog-description">
-                        Click ‘Clear History’ to delete your {dialogType}{" "}
-                        history permanently
-                    </DialogContentText>
-                </DialogContent>
-                <DialogActions>
-                    <DialogCancelButton onClick={handleClose}>
-                        Cancel
-                    </DialogCancelButton>
-                    <DialogConfirmButton onClick={handleClearHistory} autoFocus>
-                        Clear History
-                    </DialogConfirmButton>
-                </DialogActions>
-            </Dialog>
-        </div>
-    );
-}
-
-const blue = {
-    50: "#F0F7FF",
-    100: "#C2E0FF",
-    200: "#80BFFF",
-    300: "#66B2FF",
-    400: "#3399FF",
-    500: "#262626",
-    600: "#0072E5",
-    700: "#0059B2",
-    800: "#004C99",
-    900: "#003A75",
-};
-
-const Tab = styled(TabUnstyled)`
-    color: #aaa;
-    cursor: pointer;
-    font-size: 13px;
-    background-color: transparent;
-    padding: 6px 10px;
-    border: none;
-    border-radius: 3px 3px 0px 0px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-bottom: 1px solid transparent;
-    transition: 0.2s all;
-
-    &:hover {
-        background-color: #666666;
-    }
-
-    &:focus {
-        color: #aaa;
-        border-radius: 3px 3px 0px 0px;
-
-        outline-offset: 2px;
-    }
-
-    &.${tabUnstyledClasses.selected} {
-        color: #eee;
-        border-bottom: 1px solid #11abab;
-    }
-
-    &.${buttonUnstyledClasses.disabled} {
-        opacity: 0.5;
-        cursor: not-allowed;
-    }
-`;
-
-const TabHistoryIcon = styled(HistoryIcon)`
-    height: 16px;
-    width: 16px;
-    margin-right: 3px;
-`;
-
-const TabHistoryStarIcon = styled(StarBorderIcon)`
-    height: 16px;
-    width: 16px;
-    margin-right: 3px;
-`;
-const TabHistorySettingIcon = styled(DisplaySettingsIcon)`
-    height: 16px;
-    width: 16px;
-    margin-right: 3px;
-`;
-
-const TabHistoryLinkIcon = styled(LinkIcon)`
-    height: 16px;
-    width: 16px;
-    margin-right: 3px;
-`;
-
-const TabHistorySearchIcon = styled(SearchIcon)`
-    height: 21px;
-    width: 16px;
-    padding: 0px 3px;
-    border-radius: 3px 0px 0px 3px;
-    background: #121212;
-`;
-
-const TabHeaderContainer = styled.div`
-    padding: 0px 15px;
-    font-size: 13px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background: #8a8a8a50;
-    height: 37px;
-`;
-const TabPanel = styled(TabPanelUnstyled)`
-    width: 100%;
-`;
-
-const TabsList = styled(TabsListUnstyled)`
-    min-width: 320px;
-    background-color: ${blue[500]};
-    border-bottom: 4px solid #2e2e2e;
-    display: flex;
-    align-items: center;
-    align-content: space-between;
-`;
-
-const EmptyHistory = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: #ddd;
-    font-size: 14px;
-    flex: 1;
-    padding:20px;
-    height: 50%;
-`;
-function EmptyHistoryDisplay({ message }) {
-    return <EmptyHistory>{message}</EmptyHistory>;
-}
+import AlertDialog from "./components/AlertDialog/AlertDialog";
+import EmptyHistoryDisplay from "./components/EmptyHistoryDisplay/EmptyHistoryDisplay";
+import CloseButton from "./components/CloseButton/CloseButton";
 
 function QueryHistoryTabs({
     historyTabHeader,
@@ -258,183 +105,6 @@ function QueryHistoryTabs({
         </TabsUnstyled>
     );
 }
-
-export const theme = createTheme({
-    palette: {
-        mode: "dark",
-        primary: {
-            main: "#ddd",
-            background: "#1a1a1a",
-        },
-    },
-});
-
-const QueryHistoryContainer = styled.div`
-    height: 250px;
-    overflow-y: auto;
-    &.starredCont {
-        height: 210px;
-    }
-    &::-webkit-scrollbar {
-        width: 10px;
-        background: black;
-    }
-
-    &::-webkit-scrollbar-thumb {
-        border-radius: 10px;
-        background: #444;
-    }
-`;
-
-const HistoryButton = styled.button`
-    padding: 3px 6px;
-    background: #333;
-    border-radius: 3px;
-    border: none;
-    color: #ddd;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin: 0px 6px;
-    cursor: pointer;
-    min-height: 20px;
-`;
-
-const SettingItemContainer = styled.div`
-    height: 100px;
-    width: 240px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    padding: 20px;
-    background: #333;
-    margin: 10px;
-    border-radius: 3px;
-    & div {
-        font-size: 15px;
-        color: orange;
-    }
-    & small {
-        font-size: 12px;
-        color: #ddd;
-    }
-`;
-const SubmitButton = styled(HistoryButton)`
-    background: #11abab;
-    white-space: nowrap;
-    .open-icon {
-        display: none;
-    }
-    .open-text {
-        display: flex;
-    }
-    @media screen and (max-width: 864px) {
-        .open-icon {
-            display: flex;
-        }
-        .open-text {
-            display: none;
-        }
-    }
-`;
-
-const ClearHistoryButton = styled(HistoryButton)`
-    font-weight: bold;
-    padding: 10px 20px;
-    background: #088789;
-    margin: 0;
-    width: 100%;
-    white-space: nowrap;
-`;
-const StyledCloseButton = styled(HistoryButton)`
-    background: none;
-    color: #ddd;
-    font-size: 14px;
-    height: 16px;
-    width: 16px;
-    position: absolute;
-    right: 0;
-`;
-
-const DialogCancelButton = styled(HistoryButton)`
-    background: #646464;
-    padding: 8px 16px;
-`;
-const DialogConfirmButton = styled(HistoryButton)`
-    background: #088789;
-    padding: 8px 16px;
-`;
-
-const FilterInput = styled.input`
-    color: orange;
-    background: #121212;
-    border: none;
-    height: 21px;
-    margin: 0px 10px 0px 0px;
-    padding: 0px;
-    font-size: 13px;
-    border-radius: 0px 3px 3px 0px;
-    font-family: monospace;
-    font-size: 12px;
-    &:focus {
-        outline: none;
-    }
-`;
-const RowData = styled.span`
-    flex: 1;
-    font-family: "monospace";
-    font-size: "13px";
-    color: "#ddd";
-    white-space: nowrap;
-    padding: 4px 0px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-`;
-
-const LinkParams = styled.div`
-    display: flex;
-    flex: 1;
-    justify-content: space-between;
-    .open-button {
-        display: none;
-    }
-    .inline-params {
-        align-items: center;
-        display: ${(props) => (props.open ? "none" : "grid")};
-        flex: 1;
-        grid-template-columns: 1fr 0.25fr 0.25fr auto;
-        margin-right: 5px;
-    }
-
-    .open-button {
-        display: flex;
-        color: #aaa;
-        background: none;
-        border: none;
-    }
-    .block-params {
-        display: ${(props) => (props.open ? "flex" : "none")};
-        flex-direction: column;
-        flex: 1;
-        p {
-            display: flex;
-            align-items: center;
-            justify-content: space-between;
-            flex: 1;
-            border-bottom: 1px solid #333;
-            margin-bottom: 4px;
-            padding-bottom: 2px;
-            span {
-                margin-left: 3px;
-            }
-        }
-    }
-    @media screen and (max-width: 864px) {
-        .inline-params {
-            display: none;
-        }
-    }
-`;
 
 function HistoryLinkParams({
     item,
@@ -512,7 +182,15 @@ function HistoryLinkParams({
             <div className="block-params">
                 <p>
                     <span> Query:</span>{" "}
-                    <span>{decodeURIComponent(item.params.query)}</span>{" "}
+                    <span
+                        style={{
+                            margin: "5px",
+                            lineHeight: "1.5",
+                            textAlign: "right",
+                        }}
+                    >
+                        {decodeURIComponent(item.params.query)}
+                    </span>{" "}
                 </p>
                 <p>
                     <span> API URL:</span> <span>{item.params.apiUrl}</span>{" "}
@@ -600,31 +278,6 @@ function HistoryLinkTools({
     );
 }
 
-function CloseButton({ onClose }) {
-    return (
-        <StyledCloseButton onClick={onClose}>
-            <CloseIcon />
-        </StyledCloseButton>
-    );
-}
-
-const HistoryRow = styled.div`
-    padding: 5px 0px;
-    padding-left: 10px;
-    background: #212121;
-    margin: 5px;
-    border-radius: 3px;
-    font-size: 13px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    //height: 30px;
-`;
-const TimeSpan = styled.div`
-    @media screen and (max-width: 1370px) {
-        display: none;
-    }
-`;
 function QueryHistoryTab({
     queryHistory,
     copyQuery,
@@ -923,7 +576,6 @@ function SettingHistoryTabHeader({ queryHistory, linksHistory }) {
 }
 
 function SettingTab({ clearHistory, clearLinksHistory }) {
-    // const listDisplay = filtered.length > 0 ? filtered : queryHistory
     return (
         <QueryHistoryContainer>
             <div style={{ display: "flex" }}>
@@ -952,7 +604,7 @@ function SettingTab({ clearHistory, clearLinksHistory }) {
     );
 }
 
-const QueryHistoryDrawer = (props) => {
+const QueryHistory = (props) => {
     const dispatch = useDispatch();
     const historyService = localService().historyStore();
     const linkService = localUrl();
@@ -1218,10 +870,6 @@ const QueryHistoryDrawer = (props) => {
             </Drawer>
         </ThemeProvider>
     );
-};
-
-const QueryHistory = () => {
-    return <QueryHistoryDrawer />;
 };
 
 export default QueryHistory;
