+++ src/scss/modules/date-range-picker.scss
@@ -16,6 +16,7 @@
         padding: 8px;
         color:#ddd;
         line-height: 1.5;
+        margin-right:10px;
         svg {
             margin-right: 0;
         }
